requirements.txt
aiogram==2.25.1
aiogram-cache==0.1
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.10.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
backcall==0.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
contextlib2==21.6.0
cryptography==41.0.1
decorator==5.1.1
Django==4.2.2
django-extensions==3.2.3
django-ninja==0.22.2
django-ninja-extra==0.19.1
django-unixtimestampfield==0.5.0
djangorestframework==3.14.0
executing==1.2.0
frozenlist==1.3.3
idna==3.4
injector==0.20.1
ipython==8.14.0
jedi==0.18.2
magic-filter==1.0.9
matplotlib-inline==0.1.6
multidict==6.0.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.9
Pygments==2.15.1
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.6.3
tzlocal==5.0.1
urllib3==2.0.3
wcwidth==0.2.6
yarl==1.9.2

pip~=23.1.2
PyYAML~=6.0
Naked~=0.1.32
crypto~=1.4.1
shellescape~=3.8.1
psutil~=5.9.5
setuptools~=65.5.0
tomlkit~=0.11.8
mnemonic~=0.20
asn1crypto~=1.5.1
coincurve~=18.0.0
pyperclip~=1.8.2
killport~=1.2.0

beautifulsoup4~=4.12.2
soupsieve~=2.4.1
gitdb~=4.0.10
smmap~=5.0.0
passpy~=1.0.2
click~=8.1.3